Projectiles have Targets
We hired more competent soldiers so now the towers can aim at things. They're not too good at aiming, but they'll do for now.

+++ src/Path.java
@@ -17,7 +17,7 @@ public class Path extends GameObject{
 	// Draw Enemy Paths
 	public void draw(Graphics g) {
 		g.setColor(Color.RED);
-		g.fillRect(x, y, width, height);
+		g.fillRect((int)x, (int)y, width, height);
 		
 	}
 }
